+++ vantage-table/src/table/impls/table_like.rs
@@ -5,7 +5,6 @@ use vantage_expressions::AnyExpression;
 use vantage_types::Entity;
 
 use crate::{
-    any::AnyTable,
     column::flags::ColumnFlag,
     conditions::ConditionHandle,
     pagination::Pagination,
@@ -137,9 +136,5 @@ where
 
     async fn get_count(&self) -> Result<i64> {
         self.data_source().get_table_count(self).await
-    }
-
-    fn get_ref(&self, relation: &str) -> Result<AnyTable> {
-        Table::<T, E>::get_ref(self, relation)
     }
 }